[be/feat] 토큰으로 사용자 ID 획득하도록 수정

+++ backend/auction/src/main/java/com/dokidoki/auction/controller/InterestController.java
@@ -1,13 +1,15 @@
 package com.dokidoki.auction.controller;
 
 import com.dokidoki.auction.common.BaseResponseBody;
-//import com.dokidoki.auction.common.JWTutil;
+import com.dokidoki.auction.common.JWTUtil;
 import com.dokidoki.auction.service.InterestService;
 import io.swagger.annotations.ApiParam;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import javax.servlet.http.HttpServletRequest;
 
 
 @Slf4j
@@ -17,15 +19,16 @@ import org.springframework.web.bind.annotation.*;
 public class InterestController {
 
     private final InterestService interestService;
-//    private final JWTutil jwTutil;
+    private final JWTUtil jwtUtil;
 
     @PostMapping("")
     public ResponseEntity<BaseResponseBody> addInterest(
-            @RequestParam("auction_id") @ApiParam(value = "경매 id", required = true) Long auctionId
-//            HttpServletRequest request
+            @RequestParam("auction_id") @ApiParam(value = "경매 id", required = true) Long auctionId,
+            HttpServletRequest request
     ) {
-//        Long buyerId = jwTutil.getUserId(request);
-        Long buyerId = 1L;
+        Long buyerId = jwtUtil.getUserId(request);
+        if (buyerId == null)
+            return ResponseEntity.status(400).body(BaseResponseBody.of("토큰이 유효하지 않습니다."));
 
         if (interestService.addInterest(buyerId, auctionId))
             return ResponseEntity.status(201).body(BaseResponseBody.of("관심목록에 추가되었습니다."));
@@ -35,9 +38,12 @@ public class InterestController {
 
     @DeleteMapping("")
     public ResponseEntity<BaseResponseBody> deleteInterest(
-            @RequestParam("auction_id") @ApiParam(value = "경매 id", required = true) Long auctionId) {
+            @RequestParam("auction_id") @ApiParam(value = "경매 id", required = true) Long auctionId,
+            HttpServletRequest request) {
+        Long buyerId = jwtUtil.getUserId(request);
+        if (buyerId == null)
+            return ResponseEntity.status(400).body(BaseResponseBody.of("토큰이 유효하지 않습니다."));
 
-        Long buyerId = 1L;
         if (interestService.deleteInterest(buyerId, auctionId))
             return ResponseEntity.status(200).body(BaseResponseBody.of("관심 목록에서 해제되었습니다."));
         else
